+++ deploy.js
@@ -128,20 +128,9 @@ function registerCallback() {
 function install(ow) {
   WARN('Installing artifacts...');
   waterfall([
-    //   wsk package create vision
     (callback) => {
       call(ow, 'package', 'create', 'vision', callback);
     },
-    //   wsk package update vision\
-    //     -p cloudantUrl https://$CLOUDANT_username:$CLOUDANT_password@$CLOUDANT_host\
-    //     -p watsonApiKey \
-    //     -p cloudantDbName \
-    //     -p osAuthUrl "$OS_AUTH_URL"\
-    //     -p osProjectId "$OS_PROJECT_ID"\
-    //     -p osRegion "$OS_REGION"\
-    //     -p osUsername "$OS_USERNAME"\
-    //     -p osPassword "$OS_PASSWORD"\
-    //     -p osDomainId "$OS_DOMAIN_ID"
     (callback) => {
       const keyAndValues = {
         cloudantUrl: `https://${process.env.CLOUDANT_username}:${process.env.CLOUDANT_password}@${process.env.CLOUDANT_host}`,
@@ -161,6 +150,10 @@ function install(ow) {
         osUsername: process.env.OS_USERNAME || '',
         osPassword: process.env.OS_PASSWORD || '',
         osDomainId: process.env.OS_DOMAIN_ID || '',
+        cosEndpoint: process.env.COS_ENDPOINT || '',
+        cosApiKey: process.env.COS_API_KEY || '',
+        cosBucket: process.env.COS_BUCKET || '',
+        cosInstanceId: process.env.COS_INSTANCE_ID || '',
       };
       call(ow, 'package', 'update', {
         packageName: 'vision',
@@ -261,12 +254,12 @@ function makeSpeechToTextTask(ow, isCreate) {
 }
 
 function makeActionTask(ow, actionName, isCreate, options = {}) {
-  //   wsk action create vision/speechtotext --kind nodejs:6 speechtotext/speechtotext.zip
+  //   wsk action create vision/speechtotext --kind nodejs:8 speechtotext/speechtotext.zip
   return (callback) => {
     const files = {
       'package.json': `processing/${actionName}/package.json`,
       'lib/cloudantstorage.js': 'web/lib/cloudantstorage.js',
-      'lib/objectstorage.js': 'web/lib/objectstorage.js',
+      'lib/cloudobjectstorage.js': 'web/lib/cloudobjectstorage.js',
       'lib/cloudant-designs.json': 'web/lib/cloudant-designs.json'
     };
     files[`${actionName}.js`] = `processing/${actionName}/${actionName}.js`;
@@ -276,7 +269,7 @@ function makeActionTask(ow, actionName, isCreate, options = {}) {
       actionName: `vision/${actionName}`,
       action: {
         exec: {
-          kind: 'nodejs:6',
+          kind: 'nodejs:8',
           code: actionCode,
           binary: true
         },
@@ -290,7 +283,7 @@ function makeActionTask(ow, actionName, isCreate, options = {}) {
 }
 
 function makeChangeListenerTask(ow, isCreate) {
-  //   wsk action create vision-cloudant-changelistener --kind nodejs:6 changelistener/changelistener.zip\
+  //   wsk action create vision-cloudant-changelistener --kind nodejs:8 changelistener/changelistener.zip\
   //     -p cloudantUrl https://$CLOUDANT_username:$CLOUDANT_password@$CLOUDANT_host\
   //     -p cloudantDbName $CLOUDANT_db
   return (callback) => {
@@ -298,14 +291,14 @@ function makeChangeListenerTask(ow, isCreate) {
       'package.json': 'processing/changelistener/package.json',
       'changelistener.js': 'processing/changelistener/changelistener.js',
       'lib/cloudantstorage.js': 'web/lib/cloudantstorage.js',
-      'lib/objectstorage.js': 'web/lib/objectstorage.js',
+      'lib/cloudobjectstorage.js': 'web/lib/cloudobjectstorage.js',
       'lib/cloudant-designs.json': 'web/lib/cloudant-designs.json'
     });
     call(ow, 'action', isCreate ? 'create' : 'update', {
       actionName: 'vision-cloudant-changelistener',
       action: {
         exec: {
-          kind: 'nodejs:6',
+          kind: 'nodejs:8',
           code: actionCode,
           binary: true
         },