+++ processing/extractor/client/extract.js
@@ -72,29 +72,23 @@ const inputFilename = `${workingDirectory.name}/video.mp4`;
 async.waterfall([
   // establish the storage connection if configured in payload
   function(callback) {
-    if (!payload.osPassword) {
-      console.log('Media files are stored in Cloudant.');
-      callback(null, null);
-    } else {
-      const osConfig = {
-        provider: 'openstack',
-        useServiceCatalog: true,
-        useInternal: false,
-        keystoneAuthVersion: 'v3',
-        authUrl: payload.osAuthUrl,
-        tenantId: payload.osProjectId,
-        domainId: payload.osDomainId,
-        username: payload.osUsername,
-        password: payload.osPassword,
-        region: payload.osRegion
-      };
-      require('./lib/objectstorage')(osConfig, (err, fileStore) => {
+    if (payload.cosApiKey) {
+      require('./lib/cloudobjectstorage')({
+        endpoint: payload.cosEndpoint,
+        apikey: payload.cosApiKey,
+        instanceId: payload.cosInstanceId,
+        bucket: payload.cosBucket,
+      }, (err, fileStore) => {
         if (err) {
           callback(err);
         } else {
           callback(null, fileStore);
         }
       });
+      console.log('Media files are stored in Cloud Object Storage.');
+    } else {
+      console.log('Media files are stored in Cloudant.');
+      callback(null, null);
     }
   },
   // connect to the database